Move the frontend into feature folders
components/ had become a flat bag of 18 files, and pages/ and lib/ were
catch-alls. Files now live with the feature they belong to: board/,
issues/ (with issues/detail/ for the panel's sections), cycles/,
search/, imports/, team/, auth/, app/ for the shell, ui/ for the two
genuinely shared primitives. markdown/, reports/, keyboard/ and api/
were already coherent and stay put.

A @/ alias replaces relative imports throughout. An import that does
not encode its own depth survives the file being moved, which is the
property this change needed and the next one will keep needing.

Moved with git mv so history follows each file. Imports were rewritten
by resolving each one against the file's old location and mapping the
target through the move table, rather than by hand.

Nothing that ships changed: the built chunks have the same
content-addressed hashes as before the move.
