+++ frontend/src/app/main.tsx
@@ -3,9 +3,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 
-import App from './App.tsx'
-import { AuthProvider } from './auth/AuthContext'
-import './index.css'
+import App from '@/app/App'
+import { AuthProvider } from '@/auth/AuthContext'
+import '@/index.css'
 
 const queryClient = new QueryClient({
   defaultOptions: {
