+++ frontend/src/auth/RequireAuth.tsx
@@ -1,6 +1,6 @@
 import { Navigate, Outlet, useLocation } from 'react-router-dom'
 
-import { useAuth } from './AuthContext'
+import { useAuth } from '@/auth/AuthContext'
 
 export function RequireAuth() {
   const { isAuthenticated, isLoading } = useAuth()
