+++ frontend/src/board/BoardPage.tsx
@@ -6,30 +6,30 @@ import {
   getListIssuesTeamsTeamIdIssuesGetQueryKey,
   useListIssuesTeamsTeamIdIssuesGet,
   useUpdateIssueIssuesIssueIdPatch,
-} from '../api/generated/endpoints/issues/issues'
-import { useListLabelsTeamsTeamIdLabelsGet } from '../api/generated/endpoints/labels/labels'
-import { useListProjectsTeamsTeamIdProjectsGet } from '../api/generated/endpoints/projects/projects'
-import { useGetEstimateSummaryTeamsTeamIdEstimatesGet } from '../api/generated/endpoints/issues/issues'
-import { useSearchSearchGet } from '../api/generated/endpoints/search/search'
-import { useListTeamMembersTeamsTeamIdMembersGet } from '../api/generated/endpoints/teams/teams'
-import type { IssuePriority, IssueRead, IssueStatus } from '../api/generated/models'
-import { IssueDetailPanel } from '../components/IssueDetailPanel'
-import { IssueListView } from '../components/IssueListView'
-import { KanbanBoard } from '../components/KanbanBoard'
-import { CycleBanner } from '../components/CycleBanner'
-import { ImportJiraModal } from '../components/ImportJiraModal'
-import { NewCycleModal } from '../components/NewCycleModal'
-import { SearchResults } from '../components/SearchResults'
-import { ReportsView } from '../reports/ReportsView'
-import { useListCyclesTeamsTeamIdCyclesGet } from '../api/generated/endpoints/cycles/cycles'
-import { CommandPalette, type Command } from '../keyboard/CommandPalette'
-import { ShortcutsCheatsheet } from '../keyboard/ShortcutsCheatsheet'
-import { isPlainKey, isTypingTarget } from '../keyboard/typing'
-import { NewIssueModal } from '../components/NewIssueModal'
-import { Sidebar } from '../components/Sidebar'
-import { TopBar, type AssigneeFilter } from '../components/TopBar'
-import { useTeamByKey } from '../lib/useTeams'
-import { TeamProvider } from '../team/TeamContext'
+} from '@/api/generated/endpoints/issues/issues'
+import { useListLabelsTeamsTeamIdLabelsGet } from '@/api/generated/endpoints/labels/labels'
+import { useListProjectsTeamsTeamIdProjectsGet } from '@/api/generated/endpoints/projects/projects'
+import { useGetEstimateSummaryTeamsTeamIdEstimatesGet } from '@/api/generated/endpoints/issues/issues'
+import { useSearchSearchGet } from '@/api/generated/endpoints/search/search'
+import { useListTeamMembersTeamsTeamIdMembersGet } from '@/api/generated/endpoints/teams/teams'
+import type { IssuePriority, IssueRead, IssueStatus } from '@/api/generated/models'
+import { IssueDetailPanel } from '@/issues/IssueDetailPanel'
+import { IssueListView } from '@/board/IssueListView'
+import { KanbanBoard } from '@/board/KanbanBoard'
+import { CycleBanner } from '@/cycles/CycleBanner'
+import { ImportJiraModal } from '@/imports/ImportJiraModal'
+import { NewCycleModal } from '@/cycles/NewCycleModal'
+import { SearchResults } from '@/search/SearchResults'
+import { ReportsView } from '@/reports/ReportsView'
+import { useListCyclesTeamsTeamIdCyclesGet } from '@/api/generated/endpoints/cycles/cycles'
+import { CommandPalette, type Command } from '@/keyboard/CommandPalette'
+import { ShortcutsCheatsheet } from '@/keyboard/ShortcutsCheatsheet'
+import { isPlainKey, isTypingTarget } from '@/keyboard/typing'
+import { NewIssueModal } from '@/issues/NewIssueModal'
+import { Sidebar } from '@/board/Sidebar'
+import { TopBar, type AssigneeFilter } from '@/board/TopBar'
+import { useTeamByKey } from '@/team/useTeams'
+import { TeamProvider } from '@/team/TeamContext'
 
 export default function BoardPage() {
   const { teamKey, issueNumber } = useParams<{ teamKey: string; issueNumber?: string }>()
