+++ frontend/src/issues/EstimateBadge.tsx
