+++ frontend/src/markdown/Markdown.tsx
@@ -3,8 +3,8 @@ import ReactMarkdown, { type Components } from 'react-markdown'
 import type { PluggableList } from 'unified'
 import remarkGfm from 'remark-gfm'
 
-import type { Mentionable } from './mentions'
-import { remarkMentions } from './remarkMentions'
+import type { Mentionable } from '@/markdown/mentions'
+import { remarkMentions } from '@/markdown/remarkMentions'
 
 /**
  * Rendered markdown.
